brain/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
pydantic>=2.6.0
instructor>=1.3.0
openai>=1.14.0
httpx>=0.27.0
PyJWT[crypto]>=2.8.0
cryptography>=42.0.0
supabase>=2.4.0
python-dotenv>=1.0.0
